+++ src/Itq/Common/Plugin/CriteriumType/Collection/NotInDecimalCollectionCriteriumType.php
@@ -31,7 +31,7 @@ class NotInDecimalCollectionCriteriumType extends Base\AbstractCollectionCriteri
                 '+not_in_double' => [
                     $k => array_map(
                         function ($vv) {
-                            return (double) $vv;
+                            return (float) $vv;
                         },
                         explode(',', $v)
                     ),
