+++ src/Itq/Common/Plugin/CriteriumType/Mongo/EqualMongoCriteriumType.php
@@ -26,7 +26,7 @@ class EqualMongoCriteriumType extends Base\AbstractMongoCriteriumType
     public function build($v, $k, array $options = [])
     {
         return [
-            ['$eq' => (double) $v],
+            ['$eq' => (float) $v],
         ];
     }
 }
